Add In Place Editing Option
Adds the --in-place command line option to save back edits to the input
file without needing to name all the input files twice using --output.

We had to restructure the CLI file flow quite a bit because we don't
want to touch the output file if running in-place and if the contents
haven't changed. (And because argparse.FileType is now deprecated.)

To prove that the file hasn't changed we keep the input string around,
but this will cost considerable memory for large files. Potential
alternatives would be to SHA the file contents at the cost of
computation cycles and the small risk of collision, or to write out to a
temporary file and do the compare before swapping.